backend/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.3
boto3==1.16.11
botocore==1.19.11
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.2
django-allauth==0.42.0
django-cors-headers==3.5.0
django-countries==6.1.3
django-rest-auth==0.9.5
django-storages==1.10.1
djangorestframework==3.12.1
EZGmail==2020.9.29
filelock==3.0.12
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
jmespath==0.10.0
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
stripe==2.54.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.33
whitenoise==5.2.0
wrapt==1.11.2
